* Banshee/MiragePlugin.cs: Fix a crash when a file not analyzed was   deleted.
git-svn-id: https://svn.hop.at/repos/mirage@76 3ca7edfc-973d-4f12-98f5-7359feb43fe1

+++ Banshee/MiragePlugin.cs
@@ -373,15 +373,18 @@ namespace Banshee.Mirage
             while(reader.Read()) {
                 track_ids.Add (Convert.ToInt32(reader["TrackID"]));
             }
-            db.RemoveTracks (track_ids.ToArray());
             
-            StringBuilder removeSql = new StringBuilder ("DELETE FROM MirageProcessed WHERE TrackID IN (");
-            removeSql.Append (track_ids[0].ToString());
-            for (int i = 1; i < track_ids.Count; i++) {
-                removeSql.AppendFormat(", {0}", track_ids[i]);
+            if (track_ids.Count > 0) {
+                db.RemoveTracks (track_ids.ToArray());
+                
+                StringBuilder removeSql = new StringBuilder ("DELETE FROM MirageProcessed WHERE TrackID IN (");
+                removeSql.Append (track_ids[0].ToString());
+                for (int i = 1; i < track_ids.Count; i++) {
+                    removeSql.AppendFormat(", {0}", track_ids[i]);
+                }
+                removeSql.Append (")");
+                ServiceManager.DbConnection.Execute (removeSql.ToString());
             }
-            removeSql.Append (")");
-            ServiceManager.DbConnection.Execute (removeSql.ToString());
         }
         
         private void ResetMirageProcessed()
